Move verify dataset shares task to datasharing module
