+++ backend/dataall/modules/datasets/api/dataset/mutations.py
@@ -11,7 +11,6 @@ from dataall.modules.datasets.api.dataset.resolvers import (
     delete_dataset,
     import_dataset,
     start_crawler,
-    verify_dataset_share_objects,
 )
 
 createDataset = gql.MutationField(
@@ -68,11 +67,4 @@ StartGlueCrawler = gql.MutationField(
     ],
     resolver=start_crawler,
     type=gql.Ref('GlueCrawler'),
-)
-
-verifyDatasetShareObjects = gql.MutationField(
-    name='verifyDatasetShareObjects',
-    args=[gql.Argument(name='input', type=gql.Ref('ShareObjectSelectorInput'))],
-    type=gql.Boolean,
-    resolver=verify_dataset_share_objects,
 )
